requirements.txt
email-validator>=2.2.0
flask-login>=0.6.3
flask>=3.1.1
flask-cors>=6.0.1
flask-dance>=7.1.0
flask-migrate>=4.0.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
openai>=1.78.1
psycopg2-binary>=2.9.10
stripe>=12.1.0
trafilatura>=2.0.0
oauthlib>=3.2.2
pyjwt>=2.10.1
python-dotenv>=1.1.1
reportlab>=4.0.0